+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -14,6 +14,7 @@ group: Deployment & Configs
 
 - Added 'Element Geometry References' component.
 - Added 'Element Visibility' component.
+- Added 'Elevation Filter' component.
 
 {% endcapture %}
 {% include ltr/release_header_next.html title="Upcoming Changes" note=rc_release_notes %}